pyproject.toml
[project]
name = "tana-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "pyperclip>=1.11.0",
    "pytest>=9.0.1",
    "requests>=2.32.3",
    "things3-api>=2.7.0",
]
